Refs #2685 -- Isolated per-release docs via a separate process to avoid Sphinx module caching.
`update_docs` builds every release (across all supported Django versions
and languages) in a single long-running process. Each checkout's
docs/_ext/djangodocs.py is loaded by dynamically appending that
checkout's own docs/_ext directory to sys.path and importing it by
bare module name ("djangodocs"). Once Python caches that name in
sys.modules, a later checkout's import of the same name silently
reused whichever checkout's module got imported first in the process,
regardless of which checkout was actually being built.

In practice this meant a newer branch's djangodocs.py (which no longer
registers the "djangohtml" builder) got reused for older branches that
still require it, so those older branches failed outright with
"Builder name djangohtml not registered", and worse, that failure
aborted publishing for that release entirely -- including versions
whose json build had already succeeded.

Split the actual Sphinx build out into a new build_doc_release
management command, invoked by `update_docs` as a subprocess, one
process per release. `update_docs` now only handles orchestration: which
releases need building, syncing their git checkouts (shared across a
version's languages) and translations, and dispatching the build.
build_doc_release does nothing but build an already-synced checkout,
in its own fresh interpreter, so no Python state can leak between
checkouts regardless of build order.
